--- a/edb/edgeql/compiler/casts.py
+++ b/edb/edgeql/compiler/casts.py
@@ -29,7 +29,6 @@ from edb import errors
 from edb.common import parsing
 
 from edb.ir import ast as irast
-from edb.ir import typeutils as irtyputils
 from edb.ir import utils as irutils
 
 from edb.schema import casts as s_casts
@@ -71,9 +70,7 @@ def compile_cast(
 
     elif irutils.is_untyped_empty_array_expr(ir_expr):
         # Ditto for empty arrays.
-        new_typeref = irtyputils.type_to_typeref(
-            ctx.env.schema, new_stype, cache=ctx.env.type_ref_cache
-        )
+        new_typeref = typegen.type_to_typeref(new_stype, ctx.env)
         return setgen.ensure_set(
             irast.Array(elements=[], typeref=new_typeref), ctx=ctx)
 
@@ -163,12 +160,8 @@ def _cast_to_ir(
         new_stype: s_types.Type, *,
         ctx: context.ContextLevel) -> irast.Set:
 
-    orig_typeref = irtyputils.type_to_typeref(
-        ctx.env.schema, orig_stype, cache=ctx.env.type_ref_cache
-    )
-    new_typeref = irtyputils.type_to_typeref(
-        ctx.env.schema, new_stype, cache=ctx.env.type_ref_cache,
-    )
+    orig_typeref = typegen.type_to_typeref(orig_stype, env=ctx.env)
+    new_typeref = typegen.type_to_typeref(new_stype, env=ctx.env)
     cast_name = cast.get_name(ctx.env.schema)
     cast_ir = irast.TypeCast(
         expr=ir_set,
@@ -191,12 +184,8 @@ def _inheritance_cast_to_ir(
         new_stype: s_types.Type, *,
         ctx: context.ContextLevel) -> irast.Set:
 
-    orig_typeref = irtyputils.type_to_typeref(
-        ctx.env.schema, orig_stype, cache=ctx.env.type_ref_cache,
-    )
-    new_typeref = irtyputils.type_to_typeref(
-        ctx.env.schema, new_stype, cache=ctx.env.type_ref_cache,
-    )
+    orig_typeref = typegen.type_to_typeref(orig_stype, env=ctx.env)
+    new_typeref = typegen.type_to_typeref(new_stype, env=ctx.env)
     cast_ir = irast.TypeCast(
         expr=ir_set,
         from_type=orig_typeref,
@@ -495,13 +484,8 @@ def _cast_array_literal(
 
     assert isinstance(ir_set.expr, irast.Array)
 
-    orig_typeref = irtyputils.type_to_typeref(
-        ctx.env.schema, orig_stype, cache=ctx.env.type_ref_cache,
-    )
-    new_typeref = irtyputils.type_to_typeref(
-        ctx.env.schema, new_stype, cache=ctx.env.type_ref_cache,
-    )
-
+    orig_typeref = typegen.type_to_typeref(orig_stype, env=ctx.env)
+    new_typeref = typegen.type_to_typeref(new_stype, env=ctx.env)
     direct_cast = _find_cast(orig_stype, new_stype, srcctx=srcctx, ctx=ctx)
 
     if direct_cast is None: