--- a/edb/edgeql/compiler/pathctx.py
+++ b/edb/edgeql/compiler/pathctx.py
@@ -62,8 +62,9 @@ def get_tuple_indirection_path_id(
     ptrref = irtyputils.ptrref_from_ptrcls(
         schema=ctx.env.schema,
         ptrcls=ptrcls,
-        cache=ctx.env.ptr_ref_cache,
-        typeref_cache=ctx.env.type_ref_cache,
+        # FIXME: caching disabled here since it breaks tests
+        # cache=ctx.env.ptr_ref_cache,
+        # typeref_cache=ctx.env.type_ref_cache,
     )
 
     return tuple_path_id.extend(schema=ctx.env.schema, ptrref=ptrref)