Before:
```
compile_edgeql_script: Mean +- std dev: 2.60 ms +- 0.07 ms
compile_edgeql_script_webapp_configure_system: Mean +- std dev: 44.0 ms +- 3.6 ms
compile_edgeql_script_webapp_get_movie: Mean +- std dev: 289 ms +- 55 ms
compile_edgeql_script_webapp_get_person: Mean +- std dev: 316 ms +- 73 ms
compile_edgeql_script_webapp_get_user: Mean +- std dev: 150 ms +- 4 ms
compile_edgeql_script_webapp_insert_movie: Mean +- std dev: 155 ms +- 2 ms
compile_edgeql_script_webapp_insert_person: Mean +- std dev: 59.6 ms +- 2.2 ms
compile_edgeql_script_webapp_insert_review: Mean +- std dev: 91.4 ms +- 1.7 ms
```

After:
```
compile_edgeql_script: Mean +- std dev: 2.44 ms +- 0.06 ms (6.2% faster)
compile_edgeql_script_webapp_configure_system: Mean +- std dev: 35.0 ms +- 0.9 ms (21% faster)
compile_edgeql_script_webapp_get_movie: Mean +- std dev: 256 ms +- 63 ms (11.5% faster)
compile_edgeql_script_webapp_get_person: Mean +- std dev: 325 ms +- 83 ms (2.8% SLOWER)
compile_edgeql_script_webapp_get_user: Mean +- std dev: 142 ms +- 3 ms (5.4% faster)
compile_edgeql_script_webapp_insert_movie: Mean +- std dev: 140 ms +- 3 ms (9.7% faster)
compile_edgeql_script_webapp_insert_person: Mean +- std dev: 54.6 ms +- 1.5 ms (8.4% faster)
compile_edgeql_script_webapp_insert_review: Mean +- std dev: 84.6 ms +- 2.0 ms (8.5% faster)
```

For big queries the jitter is higher due to GC activity as the cache needs to be cleaned out.

Other observations:
- only caching in the Environment didn't allow me to cache all places (see two marked spots) and got me thinking that it would be better to have a per-Schema "persistent" in-memory cache for this that would survive between compilation rounds; of course benchmarks would have to clean it up but for Real World use that would make the caches' hit rate much higher;
- I looked into why `get_person()` is actually slower now but have no answers yet;
- for benchmarks we want to disable GC when the function is being measured, given that those functions are not long-lived this should help with the jitter a lot;
- benchmarking with Wi-Fi off works better for me in terms of minimizing jitter;
- there are two failing tests due to the cache now but I'm not entirely sure if those are bugs, please look at them (test_edgeql_ir_scope_tree_19; test_edgeql_ir_scope_tree_24).