@ambv 2019-12-23
I did more measuring and sleuthing on this.

Stats from all unit tests:
- 799k calls
- 83.47% hit rate
- 0.000096% hit rate for types with custom typenames
- 0.000376% name hint mismatch (see below)
- 2.7% not cached at all (recursive calls for tuple subtypes)

### Conclusion 1
Due to the extremely low hit rate on types with custom typenames, I removed the custom typename from the cache key, making it smaller and thus faster for the common case.

### Conclusion 2
The two tests that are failing are failing because:
- while the type in question is properly recognized and cached;
- the schema changed during compilation (saved a bunch of derived types);
- and so the *name hint* is different.

Since the name hint is only used in annotations, and both name hints properly describe the type anyway, it is an acceptable compromise to allow for the name hints to not be derived types. I added an additional check for now to ensure that the name hints are correct every time. My intuition was that it's not worth it because the name hint mismatch rate is extremely low while the check cost would be paid on all cache hits. But it looks like it doesn't really make the cache perform much worse.

### More benchmarks
#### Before
```
compile_edgeql_script_simple: Mean +- std dev: 2.60 ms +- 0.07 ms
compile_edgeql_script_webapp_configure_system: Mean +- std dev: 44.0 ms +- 3.6 ms
compile_edgeql_script_webapp_get_movie: Mean +- std dev: 289 ms +- 55 ms
compile_edgeql_script_webapp_get_person: Mean +- std dev: 316 ms +- 73 ms
compile_edgeql_script_webapp_get_user: Mean +- std dev: 150 ms +- 4 ms
compile_edgeql_script_webapp_insert_movie: Mean +- std dev: 155 ms +- 2 ms
compile_edgeql_script_webapp_insert_person: Mean +- std dev: 59.6 ms +- 2.2 ms
compile_edgeql_script_webapp_insert_review: Mean +- std dev: 91.4 ms +- 1.7 ms
```

#### After (without name hint correction)
```
compile_edgeql_script_simple: Mean +- std dev: 2.43 ms +- 0.07 ms (6.54% faster)
compile_edgeql_script_webapp_configure_system: Mean +- std dev: 34.9 ms +- 1.0 ms (20.68% faster)
compile_edgeql_script_webapp_get_movie: Mean +- std dev: 257 ms +- 68 ms (11.07% faster)
compile_edgeql_script_webapp_get_person: Mean +- std dev: 315 ms +- 79 ms (0.32% faster)
compile_edgeql_script_webapp_get_user: Mean +- std dev: 142 ms +- 4 ms (5.33% faster)
compile_edgeql_script_webapp_insert_movie: Mean +- std dev: 143 ms +- 3 ms (7.74% faster)
compile_edgeql_script_webapp_insert_person: Mean +- std dev: 54.3 ms +- 1.1 ms (8.89% faster)
compile_edgeql_script_webapp_insert_review: Mean +- std dev: 83.8 ms +- 1.7 ms (8.32% faster)
```

#### After (with name hint correction)
```
compile_edgeql_script_simple: Mean +- std dev: 2.52 ms +- 0.11 ms (3.08% faster)
compile_edgeql_script_webapp_configure_system: Mean +- std dev: 34.9 ms +- 0.7 ms (20.68% faster)
compile_edgeql_script_webapp_get_movie: Mean +- std dev: 250 ms +- 58 ms (13.49% faster)
compile_edgeql_script_webapp_get_person: Mean +- std dev: 323 ms +- 84 ms (2.22% SLOWER)
compile_edgeql_script_webapp_get_user: Mean +- std dev: 143 ms +- 5 ms (4.67% faster)
compile_edgeql_script_webapp_insert_movie: Mean +- std dev: 143 ms +- 3 ms (7.74% faster)
compile_edgeql_script_webapp_insert_person: Mean +- std dev: 55.2 ms +- 2.1 ms (7.38% faster)
compile_edgeql_script_webapp_insert_review: Mean +- std dev: 85.3 ms +- 2.0 ms (6.67% faster)
```